+++ build.gradle
@@ -17,12 +17,7 @@ dependencies {
     implementation "org.junit.jupiter:junit-jupiter-params:${junit_jupiter_params_version}"
     implementation "org.junit.jupiter:junit-jupiter-engine:${junit_jupiter_engine_version}"
     implementation group: 'ch.qos.logback', name: 'logback-classic', version: "${logback_version}"
-    implementation 'com.google.code.gson:gson:2.8.6'
-
-
-
-
-
+    implementation "com.google.code.gson:gson:${gson_version}"
 }
 
 test{
