+++ gradle.properties
@@ -4,3 +4,4 @@ junit_jupiter_params_version=5.5.2
 junit_jupiter_api_version=5.5.2
 junit_jupiter_engine_version=5.5.2
 logback_version=1.2.3
+gson_version=2.8.6
